change package to io.github.derysudrajat.mathview add icon for math-view
