+++ app/src/main/AndroidManifest.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.derysudrajat.mathsample">
+    package="io.github.derysudrajat.mathsample">
 
-    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
         android:allowBackup="true"
@@ -11,7 +11,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.MathView">
-        <activity android:name=".mathsample.MainActivity">
+        <activity android:name="io.github.derysudrajat.mathsample.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
